search: return communities a user is member of via search_user_communities
closes #503
